Crate sapling_crypto

Modules

alt_babyjubjub
Alternative Baby Jubjub is a twisted Edwards curve defined over the BN256 scalar field, Fr. Fr modulus = 21888242871839275222246405745257275088548364400416034343698204186575808495617

baby_group_hash

baby_pedersen_hash

baby_util

babyjubjub
Baby Jubjub is a twisted Edwards curve defined over the BN256 scalar field, Fr. It takes the form 168700 * x^2 + y^2 = 1 + dx^2y^2 with d = 168696. It is birationally equivalent to a Montgomery curve of the form y^2 = x^3 + Ax^2 + x with A = 40962. This value A is the smallest integer choice such that:

circuit

constants

eddsa
This is an implementation of EdDSA as refered in literature Generation of randomness is not specified

group_hash

interpolation

jubjub
Jubjub is a twisted Edwards curve defined over the BLS12-381 scalar field, Fr. It takes the form -x^2 + y^2 = 1 + dx^2y^2 with d = -(10240/10241). It is birationally equivalent to a Montgomery curve of the form y^2 = x^3 + Ax^2 + x with A = 40962. This value A is the smallest integer choice such that:

pedersen_hash

primitives

redbabyjubjub
Implementation of RedJubjub, a specialization of RedDSA to the Jubjub curve. See section 5.4.6 of the Sapling protocol specification.

redjubjub
Implementation of RedJubjub, a specialization of RedDSA to the Jubjub curve. See section 5.4.6 of the Sapling protocol specification.

util