Module sapling_crypto::circuit

Modules

baby_ecc

baby_eddsa

baby_pedersen_hash

blake2s

boolean

ecc

float_point

lookup

multieq

multipack

num

pedersen_hash

polynomial_lookup

sapling

sha256

sprout

test

uint32

Traits

Assignment
This basically is just an extension to Option which allows for a convenient mapping to an error on None.